# src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  description: Polar Accesslink API documentation
  version: 3
  title: Polar Accesslink API
  contact:
    name: Support
    url: 'https://www.polar.com/accesslink'
    email: b2bhelpdesk@polar.com
  license:
    name: Polar Electro AccessLink Limited License Agreement
    url: 'https://www.polar.com/sites/default/files/static/ACL_Ltd_License_Agreement.pdf'
...
paths:
...
  '/v3/users/{user-id}/exercise-transactions':
    post:
      tags:
        - Training data
      summary: Create transaction
      description: >-
        Check for new training data and create a new transaction if found.
        Only data uploaded to Flow in the last 30 days will be available.
        Only data that has been uploaded to Flow after the user is registered with your client will be available.
      operationId: createExerciseTransaction
      parameters:
        - name: user-id
          in: path
          description: User identifier
          required: true
          schema:
            type: integer
            pattern: '[0-9]+'
      responses:
        '201':
          description: >-
            Return link to created transaction containing new training session data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/transaction-location'
            application/xml:
              schema:
                $ref: '#/components/schemas/transaction-location'
        '204':
          description: No content when there is no new training session data available.
        '403':
          description: User has not accepted all mandatory consents.
      security:
        - OAuth2:
            - accesslink.read_all
  '/v3/users/{user-id}/exercise-transactions/{transaction-id}':
    get:
      tags:
        - Training data
      summary: List exercises
      description: >-
        After successfully initiating a transaction, training sessions included
        within it can be retrieved with the provided transactionId.
      operationId: listExercises
      parameters:
        - name: transaction-id
          in: path
          description: Transaction identifier
          required: true
          schema:
            type: integer
            pattern: '[0-9]+'
        - name: user-id
          in: path
          description: User identifier
          required: true
          schema:
            type: integer
            pattern: '[0-9]+'
      responses:
        '200':
          description: >-
            Return list of hyperlinks to training sessions contained within the transaction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exercises'
            application/xml:
              schema:
                $ref: '#/components/schemas/exercises'
              example:
                exercises:
                  exercise:
                    - https://www.polaraccesslink.com/v3/users/12/exercise-transactions/34/exercises/56
                    - https://www.polaraccesslink.com/v3/users/12/exercise-transactions/34/exercises/120
        '204':
          description: No content when there is no data available.
        '403':
          description: User has not accepted all mandatory consents.
      security:
        - OAuth2:
            - accesslink.read_all
    put:
      tags:
        - Training data
      summary: Commit transaction
      description: >-
        After successfully retrieving training session data within a
        transaction, partners are expected to commit the transaction.
      operationId: commitExerciseTransaction
      parameters:
        - name: transaction-id
          in: path
          description: Transaction identifier
          required: true
          schema:
            type: integer
            pattern: '[0-9]+'
        - name: user-id
          in: path
          description: User identifier
          required: true
          schema:
            type: integer
            pattern: '[0-9]+'
      responses:
        '200':
          description: Transaction has been committed and data deleted.
        '204':
          description: No content when there is no data available.
        '404':
          description: No transaction was found with given transaction id.
        '403':
          description: User has not accepted all mandatory consents.
      security:
        - OAuth2:
            - accesslink.read_all
...
servers:
  - url: 'https://www.polaraccesslink.com/'
components:
  securitySchemes:
    Basic:
      type: http
      scheme: basic
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: 'https://flow.polar.com/oauth2/authorization'
          tokenUrl: 'https://polarremote.com/v2/oauth2/token'
          scopes:
            accesslink.read_all: Allows read access to user's data
  schemas:
...
    exercises:
      type: object
      properties:
        exercises:
          type: array
          description: Absolute links to individual exercises within the transaction
          items:
            type: string
          example:
            - https://www.polaraccesslink.com/v3/users/12/exercise-transactions/34/exercises/56
            - https://www.polaraccesslink.com/v3/users/12/exercise-transactions/34/exercises/120
      description: Exercise transaction container
      xml:
        name: exercises
...
    transaction-location:
      type: object
      properties:
        transaction-id:
          type: integer
          example: 122
          description: Id of the created transaction
        resource-uri:
          type: string
          format: uri
          example: >-
            https://polaraccesslink.com/v3/users/21/physical-information-transactions/32
          description: Absolute links to the created transaction
      description: Contains absolute link to the created transaction.
      xml:
        name: transaction-location
...